+++ randovania/game_connection/builder/connector_builder_factory.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import dataclasses
 
+from randovania.bitpacking.json_dataclass import JsonDataclass
 from randovania.game_connection.builder.connector_builder import ConnectorBuilder
 from randovania.game_connection.builder.debug_connector_builder import DebugConnectorBuilder
 from randovania.game_connection.builder.dolphin_connector_builder import DolphinConnectorBuilder
@@ -16,23 +17,9 @@ _CHOICE_TO_BUILDER = {
 
 
 @dataclasses.dataclass(frozen=True)
-class ConnectorBuilderOption:
+class ConnectorBuilderOption(JsonDataclass):
     choice: ConnectorBuilderChoice
     params: dict
-
-    @property
-    def as_json(self):
-        return {
-            "choice": self.choice.value,
-            "params": self.params,
-        }
-
-    @classmethod
-    def from_json(cls, value: dict) -> ConnectorBuilderOption:
-        return cls(
-            ConnectorBuilderChoice(value["choice"]),
-            value["params"],
-        )
 
     def create_builder(self) -> ConnectorBuilder:
         return _CHOICE_TO_BUILDER[self.choice](**self.params)
