+++ randovania/game_description/resources/item_resource_info.py
@@ -1,4 +1,5 @@
 import dataclasses
+from typing import NamedTuple
 
 from frozendict import frozendict
 
@@ -16,3 +17,11 @@ class ItemResourceInfo:
 
     def __str__(self):
         return self.long_name
+
+
+class InventoryItem(NamedTuple):
+    amount: int
+    capacity: int
+
+
+Inventory = dict[ItemResourceInfo, InventoryItem]
